+++ src/validator/checkers/htmlString.ts
@@ -1,6 +1,6 @@
 export default (v: any): boolean => {
   const html = new DOMParser().parseFromString(v, 'text/html');
-  const nodes = Array.from(html.body.childNodes);
+  const nodes = html.body.childNodes;
 
-  return nodes.some((node) => node.nodeType === 1);
+  return nodes ? Array.from(nodes).some((node) => node.nodeType === 1) : false;
 };